Delete and move more tests
