+++ pkg/cmdlets/git/git.go
@@ -69,12 +69,12 @@ func (r *Repo) CloneProject(owner, name, path string, out output.Outputer, an an
 		return errs.AddTips(err, tipMsg)
 	}
 
-	err = ensureCorrectProject(owner, name, filepath.Join(tempDir, constants.ConfigFileName), *project.RepoURL, out, an)
+	err = EnsureCorrectProject(owner, name, filepath.Join(tempDir, constants.ConfigFileName), *project.RepoURL, out, an)
 	if err != nil {
 		return locale.WrapError(err, "err_git_ensure_project", "Could not ensure that the activestate.yaml in the cloned repository matches the project you are activating.")
 	}
 
-	err = moveFiles(tempDir, path)
+	err = MoveFiles(tempDir, path)
 	if err != nil {
 		return locale.WrapError(err, "err_git_move_files", "Could not move cloned files")
 	}
@@ -100,7 +100,7 @@ func plainClone(path string, isBare bool, o *git.CloneOptions) (r *git.Repositor
 	return git.PlainClone(path, isBare, o)
 }
 
-func ensureCorrectProject(owner, name, projectFilePath, repoURL string, out output.Outputer, an analytics.AnalyticsDispatcher) error {
+func EnsureCorrectProject(owner, name, projectFilePath, repoURL string, out output.Outputer, an analytics.AnalyticsDispatcher) error {
 	if !fileutils.FileExists(projectFilePath) {
 		return nil
 	}
@@ -127,7 +127,7 @@ func ensureCorrectProject(owner, name, projectFilePath, repoURL string, out outp
 	return nil
 }
 
-func moveFiles(src, dest string) error {
+func MoveFiles(src, dest string) error {
 	err := verifyDestinationDirectory(dest)
 	if err != nil {
 		return locale.WrapError(err, "err_git_verify_dir", "Could not verify destination directory")
