+++ pkg/cmdlets/git/test/integration/git_test.go
@@ -1,4 +1,4 @@
-package git
+package integration
 
 import (
 	"fmt"
@@ -22,6 +22,7 @@ import (
 	authMock "github.com/ActiveState/cli/pkg/platform/authentication/mock"
 	"github.com/ActiveState/cli/pkg/project"
 	"github.com/ActiveState/cli/pkg/projectfile"
+	gitlet "github.com/ActiveState/cli/pkg/cmdlets/git"
 )
 
 type GitTestSuite struct {
@@ -89,7 +90,7 @@ func (suite *GitTestSuite) AfterTest(suiteName, testName string) {
 }
 
 func (suite *GitTestSuite) TestEnsureCorrectProject() {
-	err := ensureCorrectProject("test-owner", "test-project", filepath.Join(suite.dir, constants.ConfigFileName), "test-repo", outputhelper.NewCatcher(), analytics.New())
+	err := gitlet.EnsureCorrectProject("test-owner", "test-project", filepath.Join(suite.dir, constants.ConfigFileName), "test-repo", outputhelper.NewCatcher(), analytics.New())
 	suite.NoError(err, "projectfile URL should contain owner and name")
 }
 
@@ -98,7 +99,7 @@ func (suite *GitTestSuite) TestEnsureCorrectProject_Mistmatch() {
 	name := "bad-project"
 	projectPath := filepath.Join(suite.dir, constants.ConfigFileName)
 	actualCatcher := outputhelper.NewCatcher()
-	err := ensureCorrectProject(owner, name, projectPath, "test-repo", actualCatcher, analytics.New())
+	err := gitlet.EnsureCorrectProject(owner, name, projectPath, "test-repo", actualCatcher, analytics.New())
 	suite.NoError(err)
 
 	proj, err := project.Parse(projectPath)
@@ -114,7 +115,7 @@ func (suite *GitTestSuite) TestEnsureCorrectProject_Mistmatch() {
 
 func (suite *GitTestSuite) TestMoveFiles() {
 	anotherDir := filepath.Join(suite.anotherDir, "anotherDir")
-	err := moveFiles(suite.dir, anotherDir)
+	err := gitlet.MoveFiles(suite.dir, anotherDir)
 	suite.NoError(err, "should be able to move files wihout error")
 
 	_, err = os.Stat(filepath.Join(anotherDir, constants.ConfigFileName))
@@ -132,7 +133,7 @@ func (suite *GitTestSuite) TestMoveFilesDirNoEmpty() {
 	err = fileutils.Touch(filepath.Join(anotherDir, "file.txt"))
 	suite.Require().NoError(err)
 
-	err = moveFiles(suite.dir, anotherDir)
+	err = gitlet.MoveFiles(suite.dir, anotherDir)
 	expected := locale.WrapError(err, "err_git_verify_dir", "Could not verify destination directory")
 	suite.EqualError(err, expected.Error())
 }
